+++ wireless_examples/reference_design/lowpower_peripheral/lowpower_peripheral.c
@@ -245,6 +245,9 @@ static void BleApp_SerialRxCb( void *params, serial_manager_callback_message_t *
 static void BleApp_SetLowPowerModeConstraint(uint8_t lowpower_mode_constraint);
 static void BleApp_ReleaseLowPowerModeConstraint(uint8_t lowpower_mode_constraint);
 
+/* Callback to handle returned errors from platform files */
+static void BleApp_PlatformErrorCallback(uint32_t id, int32_t error_status);
+
 /************************************************************************************
 *************************************************************************************
 * Public functions
@@ -268,6 +271,9 @@ void BluetoothLEHost_AppInit(void)
     BUTTON_InstallCallback((button_handle_t)g_buttonHandle[1], BleApp_HandleKeys1, NULL);
 #endif /*gAppButtonCnt_c > 1*/
 #endif /*gAppButtonCnt_c > 0*/
+
+    /* Register the application callback to handle the errors returned from platform functions */
+    PLATFORM_RegisterErrorCallback(&BleApp_PlatformErrorCallback);
 
 #if defined(gAppLowpowerEnabled_d) && (gAppLowpowerEnabled_d>0)
     PLATFORM_ResetStatus_t resetStatus;
@@ -1652,6 +1658,34 @@ static void BleApp_SerialRxCb( void *params, serial_manager_callback_message_t *
     BleApp_SerialWrite("\r\n");
 }
 #endif
+
+static void BleApp_PlatformErrorCallback(uint32_t id, int32_t error_status)
+{
+    switch(id)
+    {
+    case PLATFORM_REMOTE_ACTIVE_REQ_ID:
+        {
+            /* Callback has been called from PLATFORM_RemoteActiveReq() function */
+            assert(0);
+            break;
+        }
+    case PLATFORM_INIT_BLE_ID:
+        {
+            /* Callback has been called from PLATFORM_InitBle() function */
+            assert(0);
+            break;
+        }
+    case PLATFORM_SEND_HCI_MESSAGE_ID:
+        {
+            /* Callback has been called from PLATFORM_SendHciMessage() function */
+            assert(0);
+            break;
+        }
+    }
+
+    /* Here the error_status argument is ignored but you can decide to check it in your implementation */
+    (void)error_status;
+}
 
 /*! *********************************************************************************
 * @}