+++ _boards/kw45b41zevk/CHANGELOG.md
@@ -1,3 +1,6 @@
+#### 25.06.00-pvw2
+- [Wireless]
+  - [refdes] : Register a platform error callback in reference design applications. This callback is used to handle error reported by the platform files.
 #### 24.12.00-pvw1
 - [Wireless]
   - update macro default values : BOARD_32MHZ_XTAL_CDAC_VALUE to 12, BOARD_32MHZ_XTAL_ISEL_VALUE to 5