+++ _boards/kw47loc/CHANGELOG.md
@@ -1,3 +1,6 @@
+#### 25.06.00-pvw2
+- [Wireless]
+  - [refdes] : Register a platform error callback in reference design applications. This callback is used to handle error reported by the platform files.
 #### 24.12.00 final release
 - [Wireless]
   - [refdes] : Change gAppLowPowerConstraintInNoBleActivity_c default value from deep sleep to deep power down (2 to 4)